feat: Complete A-Z CI/CD pipeline with self-hosted runners
- Configure complete A-Z pipeline that runs ALL steps every time
- Remove conditional logic to ensure comprehensive testing and deployment
- Add missing backend job to cover all components
- Update all workflows to use self-hosted runners consistently
- Fix Docker Alpine OpenSSL linking for Azure SDK compatibility
- Resolve HTTP type incompatibility errors in Rust core API
- Add development scripts for local testing and frontend-only mode

Components covered:
- Frontend CI (Next.js)
- Core CI (Rust)
- GraphQL gateway CI
- Backend services CI
- Security scanning
- Supply chain security
- Secret scanning (Gitleaks)
- Infrastructure deployment
- Application CI/CD pipeline
- Tag-based production deployment

+++ core/src/cache.rs
@@ -254,7 +254,7 @@ impl CacheManager {
             };
 
             let json_str = serde_json::to_string(&entry)?;
-            pipe.set_ex(key, json_str, ttl.as_secs() as usize);
+            pipe.set_ex(key, json_str, ttl.as_secs());
         }
 
         let _: () = pipe.query_async(&mut self.connection_pool.clone()).await?;
@@ -379,7 +379,7 @@ impl CacheManager {
             let mut conn = self.connection_pool.clone();
             match timeout(
                 Duration::from_millis(5000),
-                conn.set_ex::<&str, String, ()>(key, json_str.clone(), ttl.as_secs() as usize),
+                conn.set_ex::<&str, String, ()>(key, json_str.clone(), ttl.as_secs()),
             )
             .await
             {
